pyproject.toml
[project]
name = "whisper-cantonese"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.9.0",
    "bitsandbytes>=0.46.1",
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "flash-attn>=2.8.1",
    "huggingface-hub>=0.33.4",
    "jiwer>=4.0.0",
    "joblib>=1.5.1",
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "numpy>=1.24.0,<2.0.0",
    "opencc>=1.1.9",
    "pandas>=2.3.1",
    "peft>=0.16.0",
    "pysrt>=1.1.2",
    "python-dotenv>=1.1.1",
    "pytubefix>=9.3.0",
    "rich>=14.0.0",
    "seaborn>=0.13.2",
    "soundfile>=0.13.1",
    "tenacity>=9.1.2",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.53.2",
]
